requirements.txt
absl-py==2.4.0
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fastapi==0.129.0
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
grad-cam==1.5.5
grpcio==1.78.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kagglehub==1.0.0
kagglesdk==0.1.15
keras==3.13.2
keras-hub==0.26.0
kiwisolver==1.4.9
langchain==1.2.10
langchain-core==1.2.14
langchain-openai==1.1.10
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.2
openai==2.21.0
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
psutil==7.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.13.1
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.24
tokenizers==0.22.2
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
transformers==5.2.0
ttach==0.0.3
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
xxhash==3.6.0
zstandard==0.25.0